pyproject.toml
[tool.poetry]
name = "unofficial-livecounts-api"
version = "0.1.4"
description = "Unofficial API Livecounts.io"
authors = ["Jonas <me@ifelse.one>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
urllib3 = "^2.2.2"
latest-user-agents = "^0.0.4"
python-dotenv = "^1.0.1"
pycryptodome = "^3.20.0"
validators = "^0.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Homepage = "https://pypi.org/project/unofficial-livecounts-api"
Issues = "https://github.com/jonaskahn/unofficial-livecounts-api/issues"
Source = "https://github.com/jonaskahn/unofficial-livecounts-api"
